Replaced quiz.java with EngineeringStreamSelector.java
Implemented minor changes such as adding the escape character to quotations inside String variables, and adding a missing curly brace.

This program can now be compiled.
